==> 3_Model Selection.ipynb <==
#Basic Necessities
import numpy as np
import pandas as pd

# Warnings Handling
import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning)
warnings.filterwarnings("ignore", category=FutureWarning)

# Word Embeddings Packages
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
import gensim

#ML Models Packages
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn import svm
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from xgboost import XGBClassifier
from sklearn.naive_bayes import BernoulliNB
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import SGDClassifier
from imblearn.over_sampling import SMOTE
from sklearn.metrics import f1_score


train = pd.read_csv('data/train_clean_data.csv')
test = pd.read_csv('data/test_clean_data.csv')

# Word Embeddings

## 1) Bag-Of-Words

BOWvectorize = CountVectorizer(ngram_range=(1,2))
BOW = BOWvectorize.fit_transform(train['clean_text'])
BOW.shape

print(BOW)

## 2) Tf-Idf Vectorization


TfidfVect = TfidfVectorizer(ngram_range=(1,2))
Tfidf = TfidfVect.fit_transform(train['clean_text'])
Tfidf.shape

print(Tfidf)

## 3) Word2Vec

Tokenize_tweet = train['clean_text'].apply(lambda x: x.split())

Model_W2V = gensim.models.Word2Vec(Tokenize_tweet, size=200, 
                                   window=5,  
                                   sg=2, 
                                   hs=0,
                                   negative=10, 
                                   workers=2, 
                                   seed=34 )

Model_W2V.train(Tokenize_tweet, total_examples=len(train['clean_text']), epochs=20)

print(Model_W2V.most_similar('sad'))

# Function to convert Word2vec into an array for passing into the model
def word_vec(tokens, size):
    vec = np.zeros(size).reshape((1, size))
    count = 0
    for word in tokens:
        try:
            vec += Model_W2V[word].reshape((1, size))
            count += 1
        except KeyError:
      
            continue
    
    if count != 0:
        vec = vec/count
    return vec

Arrays = np.zeros((len(Tokenize_tweet), 200))

for i in range(len(Tokenize_tweet)):
    Arrays[i,:] = word_vec(Tokenize_tweet[i], 200)
Arr_df = pd.DataFrame(Arrays)
Arr_df.shape

# Model Selection

## Split data into train and validation set

# BOW
XTrain_Bow, XValid_Bow, YTrain, YValid = train_test_split(BOW, train['sentiment'],
                                                          random_state = 42,
                                                          test_size=0.2)

# Tf-idf 
XTrain_tfidf, XValid_tfidf, YTrain1, YValid1 = train_test_split(Tfidf, train['sentiment'],
                                                          random_state = 42,
                                                          test_size=0.2)

# Word2Vec
XTrain_w2v, XValid_w2v, YTrain2, YValid2 = train_test_split(Arr_df, train['sentiment'],
                                                          random_state = 42,
                                                          test_size=0.2)

## 1) Logistic Regression

# BOW
LR = LogisticRegression(solver='lbfgs', max_iter=500, multi_class='multinomial')
LR.fit(XTrain_Bow,YTrain)

Prediction = LR.predict(XValid_Bow)

f1_score(YValid, Prediction, average=None)

# Tf-idf
LR = LogisticRegression(solver='lbfgs', max_iter=500, multi_class='multinomial')
LR.fit(XTrain_tfidf,YTrain1)

Prediction = LR.predict(XValid_tfidf)

f1_score(YValid, Prediction, average=None)

# Word2Vec
LR = LogisticRegression(solver='lbfgs', max_iter=500, multi_class='multinomial')
LR.fit(XTrain_w2v,YTrain2)

Prediction = LR.predict(XValid_w2v)

f1_score(YValid, Prediction, average=None)

## 2) Support Vector Machine (SVM)


# BOW
svc = svm.SVC(kernel='linear', C=1, probability=True)
svc.fit(XTrain_Bow, YTrain)

Prediction = svc.predict(XValid_Bow)

f1_score(YValid, Prediction, average=None)

# Tf-idf
svc.fit(XTrain_tfidf, YTrain1)

Prediction = svc.predict(XValid_tfidf)

f1_score(YValid1, Prediction, average=None)

# Word2Vec
svc.fit(XTrain_w2v, YTrain2)

Prediction = svc.predict(XValid_w2v)

f1_score(YValid2, Prediction , average=None)

## 3) Decision Tree

# BOW
dt = DecisionTreeClassifier(random_state=0, max_depth=6)

dt.fit(XTrain_Bow, YTrain)

Prediction = dt.predict(XValid_Bow)
f1_score(YValid, Prediction, average=None)

# Tf-idf
dt = DecisionTreeClassifier(random_state=0, max_depth=6)

dt.fit(XTrain_tfidf, YTrain1)

Prediction = dt.predict(XValid_tfidf)
f1_score(YValid, Prediction, average=None)

# Word2vec
dt = DecisionTreeClassifier(random_state=0, max_depth=6)

dt.fit(XTrain_w2v, YTrain2)

Prediction = dt.predict(XValid_w2v)
f1_score(YValid, Prediction, average=None)

## 4) Random Forest

# BOW
rf = RandomForestClassifier(n_estimators=400, random_state=11)

rf.fit(XTrain_Bow, YTrain)

Prediction = rf.predict(XValid_Bow)
f1_score(YValid, Prediction, average=None)

# Tf-idf
rf.fit(XTrain_tfidf, YTrain1)

Prediction = rf.predict(XValid_tfidf)
f1_score(YValid1, Prediction, average=None)

# Word2vec
rf.fit(XTrain_w2v, YTrain2)

Prediction = rf.predict(XValid_w2v)
f1_score(YValid2, Prediction, average=None)

## 5) Bernoulli Naive Bayes 

# BOW
BNB = BernoulliNB()
BNB.fit(XTrain_Bow, YTrain)
Prediction = BNB.predict(XValid_Bow)

f1_score(YValid, Prediction, average=None)

# Tf-idf
BNB = BernoulliNB()
BNB.fit(XTrain_tfidf,YTrain1)
Prediction = BNB.predict(XValid_tfidf)

f1_score(YValid1, Prediction, average=None)

# Word2Vec
BNB = BernoulliNB()
BNB.fit(XTrain_w2v,YTrain2)
Prediction = BNB.predict(XValid_w2v)

f1_score(YValid2, Prediction, average=None)

## 6) Multinomial Naive Bayes

# BOW
MNB = MultinomialNB()
MNB.fit(XTrain_Bow, YTrain)
Prediction = MNB.predict(XValid_Bow)

f1_score(YValid, Prediction, average=None)

# Tf-idf
MNB = MultinomialNB()
MNB.fit(XTrain_tfidf,YTrain1)
Prediction = MNB.predict(XValid_tfidf)

f1_score(YValid1, Prediction, average=None)

# Word2Vec
# Multinomial Naive Bayes cannot run on Word2Vec models
# Since, multinomial Naive Bayes doesn't accept negative values as inputs.Whereas,Word2Vec models contains negative values in it

## 7) SGD Classification

# BOW
SGD = SGDClassifier()
SGD.fit(XTrain_Bow, YTrain)
Prediction = SGD.predict(XValid_Bow)

f1_score(YValid, Prediction, average=None)

# Tf-idf
SGD = SGDClassifier()
SGD.fit(XTrain_tfidf, YTrain1)
Prediction = SGD.predict(XValid_tfidf)

f1_score(YValid1, Prediction, average=None)

# Word2Vec
SGD = SGDClassifier()
SGD.fit(XTrain_w2v, YTrain2)
Prediction = SGD.predict(XValid_w2v)

f1_score(YValid2, Prediction, average=None)

## 8) Extreme Gradient Boosting (XGBoost)

# BOW
xgbModel = XGBClassifier(Max_depth=6, n_estimators=1000)

xgbModel.fit(XTrain_Bow,YTrain)

Prediction = xgbModel.predict(XValid_Bow)
f1_score(YValid, Prediction, average=None)


# Tf-idf
xgbModel.fit(XTrain_tfidf,YTrain1)


Prediction = xgbModel.predict(XValid_tfidf)
f1_score(YValid1, Prediction, average=None)


# Word2Vec
xgbModel = XGBClassifier(Max_depth=6, n_estimators=1000, nthread=3)


xgbModel.fit(XTrain_w2v,YTrain2)

Prediction = xgbModel.predict(XValid_w2v)
f1_score(YValid2, Prediction, average=None)


# Up Sampling - Logistic Regression

# BOW
sm= SMOTE()
X,y = sm.fit_resample(XTrain_Bow,YTrain)
LR = LogisticRegression(solver='lbfgs', max_iter=500, multi_class='multinomial')
LR.fit(X,y)

Prediction = LR.predict(XValid_Bow)

f1_score(YValid, Prediction, average=None)

# Tf-idf
sm= SMOTE()
X,y = sm.fit_resample(XTrain_tfidf,YTrain1)

LR.fit(X,y)

Prediction = LR.predict(XValid_tfidf)

f1_score(YValid1, Prediction, average=None)

# Word2vec
sm= SMOTE()
X,y = sm.fit_resample(XTrain_w2v,YTrain2)

LR.fit(X,y)

Prediction = LR.predict(XValid_w2v)

f1_score(YValid2, Prediction, average=None)
